feat(Project): Redux is being added to the project, entire folder structure is changing
I'm adding 'react-redux' to the application. I've been studying the
design and architecture of Redux and it makes a lot of sense. I'll be
watching more videos and implementing it in the coming days. The
addition of Redux changed the project structure from a pure React one,
to a React/Redux structure. I'm using the ducks pattern folder
structure. I might want to consider just going pure TypeScript and using
a "features folder" approach though...

BREAKING CHANGE: None

Closes None
