+++ package.json
@@ -20,10 +20,12 @@
     "react": "^16.12.0",
     "react-apollo": "^3.1.3",
     "react-dom": "^16.12.0",
+    "react-redux": "^7.1.3",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
     "react-select": "^3.0.8",
+    "redux": "^4.0.5",
     "stylelint": "^12.0.1"
   },
   "scripts": {
