--- src/components/navigation/header/Header.module.scss
@@ -1,73 +1,0 @@
-@import '../../../styles/_variables.scss';
-
-$navbar-background-color: $body-background-color;
-
-.header {
-  background-color: $egg-white;
-}
-
-h1 {
-  /* em is relative to font-size of the ELEMENT*/
-  font-size: 4em;
-  font-weight: 700;
-  margin: 1rem 0 0 0;
-  background-color: $egg-white;
-  color: #332463;
-  font-family: Arial, fantasy;
-  text-align: center;
-}
-
-h2 {
-  text-align: center;
-}
-
-//.active {
-//  color: #8b81ae;
-//  font-weight: 600;
-//  text-shadow: 0 0 8px rgba(179, 165, 244, 1);
-//}
-
-.navbar {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin: -1.5rem 0 0 0;
-
-  &__column {
-    display: flex;
-
-    &--right {
-      @extend .navbar__column;
-      padding: 0 5rem 0 0;
-    }
-
-    &--left {
-      @extend .navbar__column;
-      padding: 0 0 0 5rem;
-    }
-  }
-
-  &__link {
-    display: flex;
-    color: #332463;
-    font-size: 2em;
-    text-decoration: none;
-    /* rem is relative to ROOT Font-size */
-    padding: 0 1rem 0 0.5rem;
-
-    &:hover {
-      color: #c2afca;
-      font-weight: 600;
-      -ms-transform: scale(1.05);
-      -webkit-transform: scale(1.05);
-      transform: scale(1.05);
-    }
-
-    &--active {
-      color: #8b81ae;
-      font-weight: 600;
-      text-shadow: 0 0 8px rgba(179, 165, 244, 1);
-    }
-  }
-
-}
