--- src/components/panel/Panel.module.scss
@@ -1,58 +1,0 @@
-@import '/../../styles/_variables.scss';
-
-.panel-headers {
-  color: #7d6d84;
-  font-size: 40px;
-  text-align: center;
-  font-weight: bold;
-}
-
-
-$panel-shadow: 0 0.5em 1em -0.125em rgba($purple, 0.3), 0 0px 0 1px rgba($purple, 0.04);
-
-.panel {
-  background-color: #F5F2FF;
-  -webkit-box-shadow: 0 0 6px 1px #332463;
-  box-shadow: 0 0 6px 1px #332463;
-  border-radius: 5px;
-  -webkit-box-direction: normal;
-  -webkit-box-orient: vertical;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-direction: column;
-  flex-direction: column;
-
-  &--account-overview {
-    @extend .panel;
-    width: fit-content;
-    margin: 3rem auto;
-  }
-
-  &--shipping {
-    @extend .panel;
-    width: fit-content;
-    margin: 3rem auto;
-    height: fit-content;
-  }
-
-  &--payments {
-     @extend .panel;
-    margin: 3rem auto;
-    min-width: 656px;
-    h2 {
-       @extend .panel-headers;
-     }
-   }
-
-  &--order-history {
-    @extend .panel;
-    margin: 3rem auto;
-    min-width: 656px;
-    h2 {
-      @extend .panel-headers;
-    }
-  }
-
-
-}
