--- src/pages/app/App.scss
@@ -1,60 +1,0 @@
-@charset "utf-8";
-
-@import '../../styles/elements';
-@import '../../styles/mixins';
-@import '../../styles/variables';
-
-
-@import '../../components/navigation/header/Header.module';
-@import '../users/information/Information.module';
-@import '../../components/navigation/footer/Footer.module';
-@import '../../components/UI/input/Input.module';
-@import '../../components/panel/Panel.module';
-@import '../../components/user/sidebar/Sidebar.module';
-
-
-@import '../../../node_modules/bulma/sass/utilities/_all.sass';
-@import '../../../node_modules/bulma/sass/base/_all.sass';
-@import '../../../node_modules/bulma/sass/elements/_all.sass';
-@import '../../../node_modules/bulma/sass/form/_all.sass';
-@import '../../../node_modules/bulma/sass/components/_all.sass';
-@import '../../../node_modules/bulma/sass/layout/_all.sass';
-@import '../../../node_modules/bulma/sass/grid/_all.sass';
-
-.App {
-  height: 100%;
-  text-align: center;
-  display: flex;
-  flex-direction: column
-}
-
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
