+++ src/views/App/App.scss
@@ -1,0 +1,60 @@
+@charset "utf-8";
+
+@import '../styles/elements';
+@import '../styles/mixins';
+@import '../styles/variables';
+
+
+@import '../components/UI/header/Header.module';
+@import '../pages/users/information/Information.module';
+@import '../components/UI/footer/Footer.module';
+@import '../components/UI/input/Input.module';
+@import '../components/panel/Panel.module';
+@import '../components/user/sidebar/Sidebar.module';
+
+
+@import '../../../node_modules/bulma/sass/utilities/all';
+@import '../../../node_modules/bulma/sass/base/all';
+@import '../../../node_modules/bulma/sass/elements/all';
+@import '../../../node_modules/bulma/sass/form/all';
+@import '../../../node_modules/bulma/sass/components/all';
+@import '../../../node_modules/bulma/sass/layout/all';
+@import '../../../node_modules/bulma/sass/grid/all';
+
+.App {
+  height: 100%;
+  text-align: center;
+  display: flex;
+  flex-direction: column
+}
+
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.App-link {
+  color: #61dafb;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
