+++ src/views/components/UI/footer/Footer.module.scss
