+++ src/views/components/forms/location-select/Location-Select.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import Select from 'react-select';
+import { LOCATION_OPTIONS } from '../../../../config/locations';
+
+const LocationSelect = (props) => {
+  return (
+    <Select
+      options={LOCATION_OPTIONS}
+      value={props.selectedState}
+      id="state"
+      name="state"
+      onChange={props.selectOnChange}
+    />
+  );
+};
+
+export default LocationSelect;
