+++ src/views/components/panel/Panel.js
