+++ src/views/pages/shop/Shop.js
