+++ src/views/styles/_variables.scss
@@ -1,0 +1,45 @@
+$purple: #332463;
+$light-purple: #8b81ae;
+$dark-purple: #161024;
+$egg-white: #f2f3f4;
+$pink-purple: #f4e0fc;
+$primary-purple: #8b81ae;
+
+
+$dark: $dark-purple;
+$light: $purple;
+$grey-light: $light-purple;
+$primary: $purple;
+$link: $primary-purple;
+$widescreen-enabled: false;
+$fullhd-enabled: false;
+
+$background: $egg-white;
+$text: $primary-purple;
+
+// Update some of Bulma's component variables
+$body-background-color: $egg-white;
+$control-border-width: 2px;
+$input-shadow: none;
+
+
+// Button Variables
+$button-color: $primary-purple;
+$button-border-color: $light-purple;
+$button-hover-border-color: $purple;
+$button-hover-color: $purple;
+
+// Navbar
+$navbar-background-color: $egg-white;
+
+// Input variables
+$input-border-color: $purple;
+$input-focus-color: $light-purple;
+
+// Columns
+$column-gap: 2rem;
+
+// Footer
+$footer-background-color: $dark-purple;
+
+
